+++ src/brushes/circle_brush.class.ts
@@ -1,6 +1,8 @@
 import { fabric } from '../../HEADER';
 import { Color } from '../color';
 import { Point } from '../point.class';
+import { Shadow } from '../shadow.class';
+import { Circle } from '../shapes/circle.class';
 import { getRandomInt } from '../util/internals';
 import { Canvas } from '../__types__';
 import { BaseBrush } from './base_brush.class';
@@ -8,7 +10,7 @@ import { BaseBrush } from './base_brush.class';
 /**
  * @todo remove transient
  */
-const { Circle, Group, Shadow } = fabric;
+const { Group } = fabric;
 
 export type CircleBrushPoint = {
   x: number;