+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [next]
 
 - chore(TS): migrate Group/ActiveSelection [#8455](https://github.com/fabricjs/fabric.js/pull/8455)
+- chore(TS): migrate Image [#8443](https://github.com/fabricjs/fabric.js/pull/8443)
+- chore(TS): migrate Shadow [#8462](https://github.com/fabricjs/fabric.js/pull/8462)
+- fix(Itext): show incorrect pointer position after scale changed
 - chore(TS): migrate text classes/mixins [#8408](https://github.com/fabricjs/fabric.js/pull/8408)
 - chore(TS): migrate Collection [#8433](https://github.com/fabricjs/fabric.js/pull/8433)
 - ci(): Simplify filestats even more [#8449](https://github.com/fabricjs/fabric.js/pull/8449)